# docs/swagger.yaml
components:
  schemas:
...
    models.CollectionCategory:
      properties:
        items:
          items:
            $ref: '#/components/schemas/models.CollectionCategoryItem'
          type: array
          uniqueItems: false
        maxTiers:
          type: integer
        name:
          type: string
        texture:
          type: string
        totalTiers:
          type: integer
      type: object
    models.CollectionCategoryItem:
      properties:
        amount:
          type: integer
        amounts:
          items:
            $ref: '#/components/schemas/models.CollectionCategoryItemAmount'
          type: array
          uniqueItems: false
        id:
          type: string
        maxTier:
          type: integer
        name:
          type: string
        texture:
          type: string
        tier:
          type: integer
        totalAmount:
          type: integer
      type: object
    models.CollectionCategoryItemAmount:
      properties:
        amount:
          type: integer
        username:
          type: string
      type: object
    models.CollectionsOutput:
      properties:
        categories:
          additionalProperties:
            $ref: '#/components/schemas/models.CollectionCategory'
          type: object
        maxedCollections:
          type: integer
        totalCollections:
          type: integer
      type: object
...
    models.CrimsonIsleDojo:
      properties:
        challenges:
          items:
            $ref: '#/components/schemas/models.CrimsonIsleDojoChallenge'
          type: array
          uniqueItems: false
        totalPoints:
          type: integer
      type: object
    models.CrimsonIsleDojoChallenge:
      properties:
        id:
          type: string
        name:
          type: string
        points:
          type: integer
        rank:
          type: string
        texture:
          type: string
        time:
          type: integer
      type: object
    models.CrimsonIsleFactions:
      properties:
        barbariansReputation:
          type: integer
        magesReputation:
          type: integer
        selectedFaction:
          type: string
      type: object
    models.CrimsonIsleKuudra:
      properties:
        tiers:
          items:
            $ref: '#/components/schemas/models.CrimsonIsleKuudraTier'
          type: array
          uniqueItems: false
        totalKills:
          type: integer
      type: object
    models.CrimsonIsleKuudraTier:
      properties:
        id:
          type: string
        kills:
          type: integer
        name:
          type: string
        texture:
          type: string
      type: object
    models.CrimsonIsleOutput:
      properties:
        dojo:
          $ref: '#/components/schemas/models.CrimsonIsleDojo'
        factions:
          $ref: '#/components/schemas/models.CrimsonIsleFactions'
        kuudra:
          $ref: '#/components/schemas/models.CrimsonIsleKuudra'
      type: object
...
    models.EmbedData:
      properties:
        bank:
          type: number
        displayName:
          type: string
        dungeons:
          $ref: '#/components/schemas/models.EmbedDataDungeons'
        game_mode:
          type: string
        joined:
          type: integer
        networth:
          $ref: '#/components/schemas/models.EmbedNetworth'
        profile_cute_name:
          type: string
        profile_id:
          type: string
        purse:
          type: number
        skills:
          $ref: '#/components/schemas/models.EmbedDataSkills'
        skyblock_level:
          type: number
        slayers:
          $ref: '#/components/schemas/models.EmbedDataSlayers'
        username:
          type: string
        uuid:
          type: string
      type: object
    models.EmbedDataDungeons:
      properties:
        classAverage:
          type: number
        classes:
          additionalProperties:
            type: integer
          type: object
        dungeoneering:
          type: number
      type: object
    models.EmbedDataSkills:
      properties:
        skillAverage:
          type: number
        skills:
          additionalProperties:
            type: integer
          type: object
      type: object
    models.EmbedDataSlayers:
      properties:
        slayers:
          additionalProperties:
            type: integer
          type: object
        xp:
          type: number
      type: object
    models.EmbedNetworth:
      properties:
        nonCosmetic:
          type: number
        normal:
          type: number
      type: object
...
    models.Minion:
      properties:
        maxTier:
          type: integer
        name:
          type: string
        texture:
          type: string
        tiers:
          items:
            type: integer
          type: array
          uniqueItems: false
      type: object
    models.MinionCategory:
      properties:
        maxedMinions:
          type: integer
        maxedTiers:
          type: integer
        minions:
          items:
            $ref: '#/components/schemas/models.Minion'
          type: array
          uniqueItems: false
        texture:
          type: string
        totalMinions:
          type: integer
        totalTiers:
          type: integer
      type: object
    models.MinionSlotsOutput:
      properties:
        bonusSlots:
          type: integer
        current:
          type: integer
        next:
          type: integer
      type: object
    models.MinionsOutput:
      properties:
        maxedMinions:
          type: integer
        maxedTiers:
          type: integer
        minions:
          additionalProperties:
            $ref: '#/components/schemas/models.MinionCategory'
          type: object
        minionsSlots:
          $ref: '#/components/schemas/models.MinionSlotsOutput'
        totalMinions:
          type: integer
        totalTiers:
          type: integer
      type: object
    models.MiscAuctions:
      properties:
        bids:
          type: number
        created:
          type: number
        fees:
          type: number
        gold_earned:
          type: number
        gold_spent:
          type: number
        highest_bid:
          type: number
        no_bids:
          type: number
        total_bought:
          additionalProperties:
            type: number
          type: object
        total_sold:
          additionalProperties:
            type: number
          type: object
        won:
          type: number
      type: object
    models.MiscDamage:
      properties:
        highest_critical_damage:
          type: number
      type: object
    models.MiscDragons:
      properties:
        deaths:
          additionalProperties:
            type: number
          type: object
        ender_crystals_destroyed:
          type: integer
        fastest_kill:
          additionalProperties:
            type: number
          type: object
        last_hits:
          additionalProperties:
            type: number
          type: object
        most_damage:
          additionalProperties:
            type: number
          type: object
      type: object
    models.MiscEndstoneProtector:
      properties:
        deaths:
          type: integer
        kills:
          type: integer
      type: object
    models.MiscEssence:
      properties:
        amount:
          type: integer
        name:
          type: string
        texture:
          type: string
      type: object
    models.MiscGifts:
      properties:
        given:
          type: integer
        received:
          type: integer
      type: object
    models.MiscKill:
      properties:
        amount:
          type: integer
        name:
          type: string
      type: object
    models.MiscKills:
      properties:
        deaths:
          items:
            $ref: '#/components/schemas/models.MiscKill'
          type: array
          uniqueItems: false
        kills:
          items:
            $ref: '#/components/schemas/models.MiscKill'
          type: array
          uniqueItems: false
        total_deaths:
          type: integer
        total_kills:
          type: integer
      type: object
    models.MiscMythologicalEvent:
      properties:
        burrows_chains_complete:
          additionalProperties:
            type: number
          type: object
        burrows_dug_combat:
          additionalProperties:
            type: number
          type: object
        burrows_dug_next:
          additionalProperties:
            type: number
          type: object
        burrows_dug_treasure:
          additionalProperties:
            type: number
          type: object
        kills:
          type: number
      type: object
    models.MiscOutput:
      properties:
        auctions:
          $ref: '#/components/schemas/models.MiscAuctions'
        claimed_items:
          additionalProperties:
            type: integer
          type: object
        damage:
          $ref: '#/components/schemas/models.MiscDamage'
        dragons:
          $ref: '#/components/schemas/models.MiscDragons'
        endstone_protector:
          $ref: '#/components/schemas/models.MiscEndstoneProtector'
        essence:
          items:
            $ref: '#/components/schemas/models.MiscEssence'
          type: array
          uniqueItems: false
        gifts:
          $ref: '#/components/schemas/models.MiscGifts'
        kills:
          $ref: '#/components/schemas/models.MiscKills'
        mythological_event:
          $ref: '#/components/schemas/models.MiscMythologicalEvent'
        pet_milestones:
          additionalProperties:
            $ref: '#/components/schemas/models.MiscPetMilestone'
          type: object
        profile_upgrades:
          $ref: '#/components/schemas/models.MiscProfileUpgrades'
        season_of_jerry:
          $ref: '#/components/schemas/models.MiscSeasonOfJerry'
        uncategorized:
          additionalProperties: {}
          type: object
      type: object
    models.MiscPetMilestone:
      properties:
        amount:
          type: integer
        progress:
          type: string
        rarity:
          type: string
        total:
          type: integer
      type: object
    models.MiscProfileUpgrades:
      additionalProperties:
        type: integer
      type: object
    models.MiscSeasonOfJerry:
      properties:
        most_cannonballs_hit:
          type: integer
        most_damage_dealt:
          type: integer
        most_magma_damage_dealt:
          type: integer
        most_snowballs_hit:
          type: integer
      type: object
...
    models.Networth:
      properties:
        nonCosmetic:
          $ref: '#/components/schemas/skycrypt_src_models.NetworthResult'
        normal:
          $ref: '#/components/schemas/skycrypt_src_models.NetworthResult'
      type: object
...
    models.ProcessingError:
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: string
      type: object
...
    skycrypt_src_models.NetworthResult:
      properties:
        bank:
          type: number
        isNonCosmetic:
          type: boolean
        networth:
          type: number
        noInventory:
          type: boolean
        personalBank:
          type: number
        purse:
          type: number
        types:
          additionalProperties:
            $ref: '#/components/schemas/skycrypt_src_models.NetworthType'
          type: object
        unsoulboundNetworth:
          type: number
      type: object
    skycrypt_src_models.NetworthType:
      properties:
        total:
          type: number
        unsoulboundTotal:
          type: number
      type: object
...
info:
  description: API for SkyCrypt - A Hypixel SkyBlock Stats Viewer
  title: SkyCrypt API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/collections/{uuid}/{profileId}:
    get:
      description: Returns collections for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.CollectionsOutput'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get collections stats of a specified player
      tags:
      - collections
  /api/crimson_isle/{uuid}/{profileId}:
    get:
      description: Returns Crimson Isle stats for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.CrimsonIsleOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get Crimson Isle stats of a specified player
      tags:
      - crimson_isle
...
  /api/embed/{uuid}:
    get:
      description: Returns embed data for the given user (UUID or username) and optional
        profile ID
      parameters:
      - description: User UUID or username
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID (optional, defaults to selected profile)
        in: query
        name: profileId
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.EmbedData'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get embed data for a specified player
      tags:
      - embed
...
  /api/item/{itemId}:
    get:
      description: Returns a PNG image of an item for the given texture ID
      parameters:
      - description: Item ID
        in: path
        name: itemId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            image/png:
              schema:
                format: binary
                type: string
          description: PNG image of the item
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Failed to render item
      summary: Render and return an item image
      tags:
      - item
...
  /api/minions/{uuid}/{profileId}:
    get:
      description: Returns minions for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.MinionsOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get minions stats of a specified player
      tags:
      - minions
  /api/misc/{uuid}/{profileId}:
    get:
      description: Returns misc stats for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.MiscOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
      summary: Get misc stats of a specified player
      tags:
      - misc
  /api/networth/{uuid}/{profileId}:
    get:
      description: Returns networth for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Networth'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get networth of a specified player
      tags:
      - networth
...
  /api/potion/{type}/{color}:
    get:
      description: Returns a PNG image of a potion for the given type and color
      parameters:
      - description: Potion Type
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: Potion Color
        in: path
        name: color
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            image/png:
              schema:
                format: binary
                type: string
          description: PNG image of the potion
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Render and return a potion image
      tags:
      - potion
...
servers:
- url: localhost:8080/